requirements.txt
requests
tweepy
fal-client
Pillow
python-dotenv
python-telegram-bot
